+++ src/core/ObservableQuery.ts
@@ -1,4 +1,8 @@
 import {
+  ExecutionResult,
+} from 'graphql';
+
+import {
   ModifiableWatchQueryOptions,
   WatchQueryOptions,
   FetchMoreQueryOptions,
@@ -272,7 +276,7 @@ export class ObservableQuery<T> extends Observable<ApolloQueryResult<T>> {
             });
         };
         this.updateQuery(mapFn);
-        return fetchMoreResult;
+        return fetchMoreResult as ApolloQueryResult<T>;
       });
   }