+++ test/optimistic.ts
@@ -741,13 +741,13 @@ describe('optimistic mutation results', () => {
       .then(() => {
         const updateQueries = {
           todoList: (prev, options) => {
-            const mResult = options.mutationResult as any;
+            const mResult = options.mutationResult;
 
             const state = cloneDeep(prev) as any;
             state.todoList.todos.unshift(mResult.data.createTodo);
             return state;
           },
-        } as MutationQueryReducersMap;
+        } as MutationQueryReducersMap<any>;
         const promise = client.mutate({
           mutation,
           optimisticResponse,
@@ -815,13 +815,13 @@ describe('optimistic mutation results', () => {
       .then(() => {
         const updateQueries = {
           todoList: (prev, options) => {
-            const mResult = options.mutationResult as any;
+            const mResult = options.mutationResult;
 
             const state = cloneDeep(prev) as any;
             state.todoList.todos.unshift(mResult.data.createTodo);
             return state;
           },
-        } as MutationQueryReducersMap;
+        } as MutationQueryReducersMap<any>;
         const promise = client.mutate({
           mutation,
           optimisticResponse,
@@ -893,13 +893,13 @@ describe('optimistic mutation results', () => {
 
       const updateQueries = {
         todoList: (prev, options) => {
-          const mResult = options.mutationResult as any;
+          const mResult = options.mutationResult;
 
           const state = cloneDeep(prev) as any;
           state.todoList.todos.unshift(mResult.data.createTodo);
           return state;
         },
-      } as MutationQueryReducersMap;
+      } as MutationQueryReducersMap<any>;
 
       client = new ApolloClient({
         networkInterface,
@@ -1689,12 +1689,12 @@ describe('optimistic mutation - githunt comments', () => {
   };
   const updateQueries = {
     Comment: (prev, { mutationResult: mutationResultArg }) => {
-      const newComment = (mutationResultArg as any).data.submitComment;
+      const newComment = mutationResultArg.data.submitComment;
       const state = cloneDeep(prev);
       (state as any).entry.comments.unshift(newComment);
       return state;
     },
-  } as MutationQueryReducersMap;
+  } as MutationQueryReducersMap<any>;
   const optimisticResponse = {
     __typename: 'Mutation',
     submitComment: {